// src/types/types.ts
export interface CurrentDate {
    day: number;
    month: number;
    year: number;
}
export interface OrdersState {
    id: string;
    name: string;
    time: string;
}

export interface UsersState {
    email: string;
    name: string;
    time: string;
    price: number;
}

export interface InputState {
    email: string;
    name: string;
    time: string;
}